n tinha ido

+++ main.py
@@ -1,7 +1,7 @@
 
-a = int(input('Digite um número a: '))
-b = int(input('Digite um numero b: '))
-c = int(input('Digite um numero c: '))
+a = float(input('Digite um número a: '))
+b = float(input('Digite um numero b: '))
+c = float(input('Digite um numero c: '))
 
 if c + b > a and a + c > b and a + b > c:
       if a == b and b == c and c == a:
@@ -11,4 +11,4 @@ if c + b > a and a + c > b and a + b > c:
       if (a == b and b != c) or (a == c and c != b) or (b == c and c != a):
             print('Isóceles')
 else:
-      print('Não é um triângulo')
+      print('Não é um triângulo')
